[login] Move classes implementation out to .cpp files
